requirements.txt
openai==1.55.3
httpx==0.27.2
adalflow==1.0.4
python-dotenv
